Remove legacy scriptlib/mssql_cli and add run_cli entrypoint
