+++ scripts/common.py
@@ -1,0 +1,76 @@
+from __future__ import annotations
+
+import importlib.util
+import os
+import types
+from datetime import datetime
+from typing import Any, Union, get_args, get_origin
+
+from pydantic import BaseModel
+
+REPO_ROOT = os.path.abspath(os.path.join(os.path.dirname(__file__), ".."))
+
+PY_TO_TS = {
+  str: "string",
+  int: "number",
+  float: "number",
+  bool: "boolean",
+  type(None): "null",
+  dict: "Record<string, any>",
+  datetime: "string",
+  list: "any[]",
+}
+
+HEADER_COMMENT = [
+  "// ================================================",
+  "// WARNING: This file is automatically generated.",
+  "// Do not modify it by hand. Your changes will be",
+  "// overwritten the next time the generator runs.",
+  "// ================================================",
+  "",
+]
+
+
+def camel_case(name: str) -> str:
+  return "".join(part.capitalize() for part in name.split("_"))
+
+
+def load_module(path: str) -> types.ModuleType:
+  module_name = os.path.splitext(os.path.basename(path))[0] + "_tmp"
+  spec = importlib.util.spec_from_file_location(module_name, path)
+  if not spec or not spec.loader:
+    raise ImportError(f"Could not load spec for module at {path}")
+  module = importlib.util.module_from_spec(spec)
+  spec.loader.exec_module(module)
+  return module
+
+
+def _py_to_ts(py_type: Any) -> str:
+  origin = get_origin(py_type)
+  args = get_args(py_type)
+  if origin in (list, tuple):
+    inner_type = _py_to_ts(args[0]) if args else "any"
+    return f"{inner_type}[]"
+  if origin is Union:
+    non_none = [arg for arg in args if arg is not type(None)]
+    if len(non_none) == 1:
+      return f"{_py_to_ts(non_none[0])} | null"
+    return " | ".join(_py_to_ts(arg) for arg in args)
+  if py_type in PY_TO_TS:
+    return PY_TO_TS[py_type]
+  if isinstance(py_type, type) and issubclass(py_type, BaseModel):
+    return py_type.__name__
+  return "any"
+
+
+def model_to_ts(model: type[BaseModel]) -> str:
+  fields = getattr(model, "model_fields", None) or getattr(model, "__fields__", {})
+  if not fields:
+    return f"export type {model.__name__} = Record<string, never>;"
+  lines = [f"export interface {model.__name__} {{"]
+  for name, field in fields.items():
+    annotation = getattr(field, "annotation", None) or getattr(field, "outer_type_", None)
+    ts_type = _py_to_ts(annotation)
+    lines.append(f"  {name}: {ts_type};")
+  lines.append("}")
+  return "\n".join(lines)
