+++ scripts/run_cli.py
@@ -1,0 +1,6 @@
+"""Entry point for the database management CLI."""
+
+from server.modules.database_cli.cli import run_repl
+
+if __name__ == "__main__":
+  run_repl()
